Make things compile (#212)
