+++ crates/world/flashblocks/Cargo.toml
@@ -9,14 +9,18 @@ license.workspace = true
 [dependencies]
 # reth
 reth.workspace = true
+reth-execution-types.workspace = true
 reth-basic-payload-builder.workspace = true
 reth-chain-state.workspace = true
+reth-chainspec.workspace = true
 reth-evm.workspace = true
 reth-provider.workspace = true
 reth-primitives.workspace = true
 reth-transaction-pool.workspace = true
 reth-primitives-traits.workspace = true
 reth-payload-util.workspace = true
+reth-payload-builder-primitives.workspace = true
+reth-payload-primitives.workspace = true
 revm.workspace = true
 
 # op-reth
@@ -30,8 +34,9 @@ reth-optimism-consensus.workspace = true
 # alloy
 alloy-consensus.workspace = true
 alloy-primitives.workspace = true
+alloy-rpc-types-eth.workspace = true
 alloy-eips.workspace = true
-alloy-serde = "0.7"
+alloy-serde.workspace = true
 
 
 # tokio
