+++ crates/world/flashblocks/src/builder.rs
@@ -1,9 +1,7 @@
-use alloy_primitives::{Bytes, U256};
 use futures_util::{sink::SinkExt, FutureExt};
 use reth::{
     api::{PayloadBuilderAttributes, PayloadBuilderError},
     chainspec::EthChainSpec,
-    payload::PayloadId,
     revm::{database::StateProviderDatabase, Database, State},
 };
 use reth_basic_payload_builder::{BuildArguments, BuildOutcome, BuildOutcomeKind};
@@ -15,25 +13,22 @@ use reth_evm::{
     ConfigureEvm, Evm,
 };
 use reth_optimism_forks::OpHardforks;
-use reth_optimism_node::OpNextBlockEnvAttributes;
+use reth_optimism_node::{txpool::interop::MaybeInteropTransaction, OpNextBlockEnvAttributes};
+use reth_optimism_payload_builder::{
+    builder::OpPayloadBuilderCtx, config::OpBuilderConfig, OpPayloadPrimitives,
+};
 use reth_optimism_payload_builder::{
     builder::OpPayloadTransactions,
     payload::{OpBuiltPayload, OpPayloadBuilderAttributes},
 };
-use reth_optimism_payload_builder::{
-    builder::{ExecutionInfo, OpPayloadBuilderCtx},
-    config::OpBuilderConfig,
-    OpPayloadPrimitives,
-};
 use reth_payload_util::{NoopPayloadTransactions, PayloadTransactions};
-use reth_primitives::{NodePrimitives, SealedHeader, TxTy};
+use reth_primitives::NodePrimitives;
 use reth_provider::{
     ChainSpecProvider, ExecutionOutcome, ProviderError, StateProvider, StateProviderFactory,
 };
 use reth_transaction_pool::{
     BestTransactionsAttributes, EthPoolTransaction, PoolTransaction, TransactionPool,
 };
-use revm::context::BlockEnv;
 use std::{
     sync::{Arc, Mutex},
     time::{Duration, Instant},
@@ -45,9 +40,11 @@ use tokio::{
 use tokio_tungstenite::{accept_async, WebSocketStream};
 use tracing::{debug, warn};
 
+use crate::ctx::PayloadBuilderCtx;
+
 /// Optimism's payload builder
 #[derive(Debug, Clone)]
-pub struct FlashBlocksPayloadBuilder<Pool, Client, Evm, Txs = ()> {
+pub struct FlashBlocksPayloadBuilder<Pool, Client, Evm, Ctx, Txs = ()> {
     /// The rollup's compute pending block configuration option.
     // TODO(clabby): Implement this feature.
     pub compute_pending_block: bool,
@@ -68,6 +65,8 @@ pub struct FlashBlocksPayloadBuilder<Pool, Client, Evm, Txs = ()> {
     pub block_time: u64,
     /// Flashblock interval in milliseconds
     pub flashblock_interval: u64,
+    /// The payload builder context
+    pub ctx: Ctx,
 }
 
 impl<Pool, Client, Evm, Txs> FlashBlocksPayloadBuilder<Pool, Client, Evm, Txs> {
@@ -148,7 +147,9 @@ where
         best: impl Fn(BestTransactionsAttributes) -> Txs + Send + Sync + 'a,
     ) -> Result<BuildOutcome<OpBuiltPayload<N>>, PayloadBuilderError>
     where
-        Txs: PayloadTransactions<Transaction: PoolTransaction<Consensus = N::SignedTx>>,
+        Txs: PayloadTransactions<
+            Transaction: MaybeInteropTransaction + PoolTransaction<Consensus = N::SignedTx>,
+        >,
     {
         let BuildArguments {
             mut cached_reads,
@@ -190,7 +191,9 @@ impl<Pool, Client, Evm, N, Txs> PayloadBuilder for FlashBlocksPayloadBuilder<Poo
 where
     Client: StateProviderFactory + ChainSpecProvider<ChainSpec: EthChainSpec + OpHardforks> + Clone,
     N: OpPayloadPrimitives,
-    Pool: TransactionPool<Transaction: PoolTransaction<Consensus = N::SignedTx>>,
+    Pool: TransactionPool<
+        Transaction: MaybeInteropTransaction + PoolTransaction<Consensus = N::SignedTx>,
+    >,
     Evm: ConfigureEvm<Primitives = N, NextBlockEnvCtx = OpNextBlockEnvAttributes>,
     Txs: OpPayloadTransactions<Pool::Transaction>,
 {
@@ -287,7 +290,9 @@ impl<Txs> FlashblockBuilder<'_, Txs> {
         EvmConfig: ConfigureEvm<Primitives = N, NextBlockEnvCtx = OpNextBlockEnvAttributes>,
         ChainSpec: EthChainSpec + OpHardforks,
         N: OpPayloadPrimitives,
-        Txs: PayloadTransactions<Transaction: PoolTransaction<Consensus = N::SignedTx>>,
+        Txs: PayloadTransactions<
+            Transaction: MaybeInteropTransaction + PoolTransaction<Consensus = N::SignedTx>,
+        >,
         Ctx: PayloadBuilderCtx<Evm = EvmConfig, ChainSpec = ChainSpec>,
     {
         debug!(target: "payload_builder", id=%ctx.payload_id(), parent_header = ?ctx.parent().hash(), parent_number = ctx.parent().number, "building new payload");
@@ -333,6 +338,8 @@ impl<Txs> FlashblockBuilder<'_, Txs> {
                 }
 
                 flashblock_gas_limit = gas_limit - info.cumulative_gas_used;
+
+                // crate::block_building::build_block(&mut builder)
 
                 // builder.finish_flashblock(&mut builder, &ctx, &mut info)?;
 
@@ -406,123 +413,4 @@ pub trait Flashblock<N: NodePrimitives> {
         self,
         state: impl StateProvider,
     ) -> Result<BlockBuilderOutcome<N>, BlockExecutionError>;
-}
-
-pub trait PayloadBuilderCtx {
-    type Evm: ConfigureEvm;
-    type ChainSpec: EthChainSpec + OpHardforks;
-
-    fn parent(&self) -> &SealedHeader;
-
-    fn attributes(
-        &self,
-    ) -> &OpPayloadBuilderAttributes<TxTy<<Self::Evm as ConfigureEvm>::Primitives>>;
-
-    fn extra_data(&self) -> Result<Bytes, PayloadBuilderError>;
-
-    fn best_transaction_attributes(&self, block_env: &BlockEnv) -> BestTransactionsAttributes;
-
-    fn payload_id(&self) -> PayloadId;
-
-    fn is_holocene_active(&self) -> bool;
-
-    fn is_better_payload(&self, total_fees: U256) -> bool;
-
-    fn block_builder<'a, DB: reth::revm::Database>(
-        &'a self,
-        db: &'a mut State<DB>,
-    ) -> Result<
-        impl BlockBuilder<Primitives = <Self::Evm as ConfigureEvm>::Primitives> + 'a,
-        PayloadBuilderError,
-    >
-    where
-        DB: Database,
-        DB::Error: Send + Sync + 'static;
-
-    fn execute_sequencer_transactions(
-        &self,
-        builder: &mut impl BlockBuilder<Primitives = <Self::Evm as ConfigureEvm>::Primitives>,
-    ) -> Result<ExecutionInfo, PayloadBuilderError>;
-
-    fn execute_best_transactions(
-        &self,
-        info: &mut ExecutionInfo,
-        builder: &mut impl BlockBuilder<Primitives = <Self::Evm as ConfigureEvm>::Primitives>,
-        best_txs: impl PayloadTransactions<
-            Transaction: PoolTransaction<Consensus = TxTy<<Self::Evm as ConfigureEvm>::Primitives>>,
-        >,
-        gas_limit: u64,
-    ) -> Result<Option<()>, PayloadBuilderError>;
-}
-
-impl<Evm, Chainspec> PayloadBuilderCtx for OpPayloadBuilderCtx<Evm, Chainspec>
-where
-    Evm: ConfigureEvm<Primitives: OpPayloadPrimitives, NextBlockEnvCtx = OpNextBlockEnvAttributes>,
-    Chainspec: EthChainSpec + OpHardforks,
-{
-    type Evm = Evm;
-    type ChainSpec = Chainspec;
-
-    fn parent(&self) -> &SealedHeader {
-        self.parent()
-    }
-
-    fn attributes(
-        &self,
-    ) -> &OpPayloadBuilderAttributes<TxTy<<Self::Evm as ConfigureEvm>::Primitives>> {
-        self.attributes()
-    }
-
-    fn extra_data(&self) -> Result<Bytes, PayloadBuilderError> {
-        self.extra_data()
-    }
-
-    fn best_transaction_attributes(&self, block_env: &BlockEnv) -> BestTransactionsAttributes {
-        self.best_transaction_attributes(block_env)
-    }
-
-    fn payload_id(&self) -> PayloadId {
-        self.payload_id()
-    }
-
-    fn is_holocene_active(&self) -> bool {
-        self.is_holocene_active()
-    }
-
-    fn is_better_payload(&self, total_fees: U256) -> bool {
-        self.is_better_payload(total_fees)
-    }
-
-    fn block_builder<'a, DB: Database>(
-        &'a self,
-        db: &'a mut State<DB>,
-    ) -> Result<
-        impl BlockBuilder<Primitives = <Self::Evm as ConfigureEvm>::Primitives> + 'a,
-        PayloadBuilderError,
-    >
-    where
-        DB: Database,
-        DB::Error: Send + Sync + 'static,
-    {
-        self.block_builder(db)
-    }
-
-    fn execute_sequencer_transactions(
-        &self,
-        builder: &mut impl BlockBuilder<Primitives = <Self::Evm as ConfigureEvm>::Primitives>,
-    ) -> Result<ExecutionInfo, PayloadBuilderError> {
-        self.execute_sequencer_transactions(builder)
-    }
-
-    fn execute_best_transactions(
-        &self,
-        info: &mut ExecutionInfo,
-        builder: &mut impl BlockBuilder<Primitives = <Self::Evm as ConfigureEvm>::Primitives>,
-        best_txs: impl PayloadTransactions<
-            Transaction: PoolTransaction<Consensus = TxTy<<Self::Evm as ConfigureEvm>::Primitives>>,
-        >,
-        _gas_limit: u64,
-    ) -> Result<Option<()>, PayloadBuilderError> {
-        self.execute_best_transactions(info, builder, best_txs)
-    }
 }
