--- world-chain-builder/crates/flashblocks/payload/src/lib.rs
@@ -1,2 +1,0 @@
-pub mod builder;
-pub mod payload;
